# staging/backend_integration_v2/documentation/api-spec.yaml
openapi: 3.0.3
info:
  title: منصة حركة - Haraka Platform API
  description: |
    واجهة برمجة التطبيقات لمنصة حركة لتحليل الأداء الحركي للطلاب
    
    ## المميزات الرئيسية
    - تحليل الفيديو بالذكاء الاصطناعي
    - إدارة ملفات الطلاب والتحليلات
    - نظام تقارير شامل
    - إشعارات ذكية في الوقت الفعلي
    
    ## المصادقة
    تستخدم المنصة JWT tokens للمصادقة. أرسل التوكن في header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## معدل الطلبات
    - 1000 طلب لكل 15 دقيقة للمستخدمين العاديين
    - 5000 طلب لكل 15 دقيقة للمديرين
    
  version: 1.0.0
  contact:
    name: فريق تطوير منصة حركة
    email: support@haraka.edu.sa
    url: https://docs.haraka.edu.sa
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.haraka.edu.sa/v1
    description: خادم الإنتاج
  - url: https://staging-api.haraka.edu.sa/v1
    description: خادم الاختبار
  - url: http://localhost:3000/api/v1
    description: خادم التطوير المحلي
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: قائمة الإشعارات
      description: الحصول على إشعارات المستخدم الحالي
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: is_read
          in: query
          schema:
            type: boolean
        - name: priority
          in: query
          schema:
            type: string
            enum: [منخفض, متوسط, عالي, عاجل, حرج]
        - name: category
          in: query
          schema:
            type: string
            enum: [عام, أداء, تنبيه, تحديث, تذكير, طارئ]
      responses:
        '200':
          description: قائمة الإشعارات
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      stats:
                        type: object
                        properties:
                          total_count:
                            type: integer
                          unread_count:
                            type: integer
  /notifications/{id}/read:
    put:
      tags:
        - Notifications
      summary: تعليم إشعار كمقروء
      description: تعليم إشعار محدد كمقروء
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: تم تعليم الإشعار كمقروء
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "تم تعليم الإشعار كمقروء"

  # Health Check
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "تنبيه: أداء منخفض"
        message:
          type: string
          example: "الطالب محمد علي حصل على نتيجة 55% وهي أقل من المطلوب"
        category:
          type: string
          enum: [عام, أداء, تنبيه, تحديث, تذكير, طارئ]
          example: "تنبيه"
        priority:
          type: string
          enum: [منخفض, متوسط, عالي, عاجل, حرج]
          example: "عالي"
        is_read:
          type: boolean
          example: false
        action_required:
          type: boolean
          example: true
        related_student_id:
          type: integer
          nullable: true
          example: 1
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
        total_count:
          type: integer
          example: 95
        has_next:
          type: boolean
          example: true
        has_previous:
          type: boolean
          example: false
...